Add timer virt (WIP)
