+++ include/sddf/timer/timer_driver.h
@@ -20,10 +20,6 @@
 
 #define LOG_TIMER_DRIVER_ERR(...) do{ sddf_dprintf("TIMER DRIVER|ERROR: "); sddf_dprintf(__VA_ARGS__); }while(0)
 
-// Includes for timer drivers; implementing a priority heap array
-// and structs for support of an arbitrary number of timeouts.
-#define SDDF_TIMER_MAX_TIMEOUTS (128)
-
 // Mult-shift cache
 typedef struct ms_cache_entry {
     sddf_timer_freq_hz_t f_a;
@@ -36,46 +32,4 @@ uint64_t tick_to_ns_cached(uint64_t ticks, uint64_t prescaler, sddf_timer_freq_h
 uint64_t ns_to_tick_cached(uint64_t ns, uint64_t prescaler, sddf_timer_freq_hz_t base_freq);
 uint64_t do_freq_shift(uint64_t t_a, uint64_t mult, uint64_t shift);
 void find_mult_shift(sddf_timer_freq_hz_t f_a, sddf_timer_freq_hz_t f_b, uint64_t *f_mult, uint64_t *f_shift);
-// Representing one timeout operation in the queue.
-typedef struct timeout {
-    uint64_t timestamp;
-    unsigned int client_channel;
-} timeout_t;
 
-// ### Timer (static) priority heap ###
-typedef struct timer_heap {
-    timeout_t timeouts[SDDF_TIMER_MAX_TIMEOUTS];
-    size_t size;
-} timer_heap_t;
-
-void timer_heap_init(timer_heap_t *heap);
-
-bool timer_heap_is_empty(const timer_heap_t *heap);
-
-bool timer_heap_is_full(const timer_heap_t *heap);
-
-size_t timer_heap_parent(size_t index);
-
-size_t timer_heap_left_child(size_t index);
-
-size_t timer_heap_right_child(size_t index);
-
-void timer_heap_swap(timer_heap_t *heap, size_t i, size_t j);
-
-// Heapify up - use after inserting
-void timer_heap_heapify_up(timer_heap_t *heap, size_t index);
-
-// Heapify down - use after popping head.
-void timer_heap_heapify_down(timer_heap_t *heap, size_t index);
-
-/**
- * Insert a new timeout into the timer priority heap.
- * @return true if successful, otherwise false
- */
-bool timer_heap_insert(timer_heap_t *heap, uint64_t timestamp, unsigned int client_channel);
-
-timeout_t *timer_heap_peek(timer_heap_t *heap);
-
-bool timer_heap_pop(timer_heap_t *heap, timeout_t *result);
-
-size_t timer_heap_size(const timer_heap_t *heap);
